Add login and logout

+++ studybud/views.py
@@ -1,9 +1,38 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
 from .models import Room, Topic
 from .forms import RoomForm
 
 # Create your views here
+def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+        
+        try:
+            user = authenticate(request, username=username, password= password)
+            if user is None:
+                raise Exception("User or password does not exist")
+            login(request, user)
+            return redirect('home')
+            
+        except Exception as e:
+            messages.error(request, e)
+
+    context = {} 
+    return render(request, 'login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -22,6 +51,7 @@ def room(request, pk):
     context = {'rooms':[room]} 
     return render(request=request, template_name='home.html', context=context)
 
+@login_required(login_url='/login')
 def create_room(request):
     if request.method == 'POST':
         #request.POST.get('name')
@@ -34,8 +64,12 @@ def create_room(request):
     context = {'form': form}
     return render(request, 'room_form.html', context)
 
+@login_required(login_url='/login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.user:
+        return HttpResponse('You are not allowed to update this room!')
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -46,9 +80,14 @@ def update_room(request, pk):
         form = RoomForm(instance=room)
     context = {'form': form}
     return render(request, 'room_form.html', context)
-    
+
+@login_required(login_url='/login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.user:
+        return HttpResponse('You are not allowed to delete this room!')
+    
     if request.method == "POST":
         room.delete()
         return redirect('home')
